- C++11 compliance fixes regarding previous commit.
git-svn-id: https://svn.linuxsampler.org/svn@2522 d2c4447f-3961-4fff-bec3-96d84be9abe5

+++ linuxsampler/trunk/src/drivers/midi/VirtualMidiDevice.h
@@ -83,7 +83,7 @@ public:
      * @returns true on success, false if internal FIFO full
      *          (or provided value invalid)
      */
-    bool SendProgramChangeToSampler(int Program);
+    bool SendProgramChangeToSampler(uint8_t Program);
 
     /**
      * Can be called by the virtual MIDI device to check whether a new note
